Add godot_scripting module
