+++ demo/.gitignore
@@ -5,4 +5,3 @@ bin/*
 !bin/*.gdextension
 python/*
 !python/.gdignore
-
